calculates lives remaining from checking all the wrong letters
From the current game session record retrived from the table, it iterates through the guesses string of letters to check which are incorrect, i.e. not present is the word_to_guess. According to this it keeps deducting from the total lives to return the remaining lives to the controller to display in the view. The controller also checks the losing condition, that is when the user has no more lives left to guess. In this case, it returns the result page with the appropriate message and the word revealed.
:wq